requirements.txt
ollama>=0.2.0
openai>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0
rich>=13.0.0
typer>=0.9.0
pyyaml>=6.0
numpy>=1.24.0
matplotlib>=3.7.0
networkx>=3.1
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
jinja2>=3.1.0